Refactor pytest fixtures
